+++ Loader/PhpFileLoader.php
@@ -98,7 +98,7 @@ class PhpFileLoader extends FileLoader
                     $result = null;
                 }
 
-                trigger_deprecation('symfony/dependency-injection', '7.4', 'Using `$this` or its internal scope in config files is deprecated, use the `$loader` variable instead in "%s" on line %d.', $e->getFile(), $e->getLine());
+                throw new LogicException(\sprintf('Using `$this` or its internal scope in config files is not supported anymore, use the `$loader` variable instead in "%s" on line %d.', $e->getFile(), $e->getLine()), $e->getCode(), $e);
             }
 
             if (\is_object($result) && \is_callable($result)) {
@@ -300,11 +300,4 @@ class PhpFileLoader extends FileLoader
 
         return $loader();
     }
-}
-
-/**
- * @internal
- */
-final class ProtectedPhpFileLoader extends PhpFileLoader
-{
 }