+++ Tests/Loader/PhpFileLoaderTest.php
@@ -331,14 +331,13 @@ class PhpFileLoaderTest extends TestCase
         $this->assertStringEqualsFile(\dirname(__DIR__).'/Fixtures/php/named_closure_compiled.php', $dumper->dump());
     }
 
-    #[IgnoreDeprecations]
-    #[Group('legacy')]
     public function testTriggersDeprecationWhenAccessingLoaderInternalScope()
     {
         $fixtures = realpath(__DIR__.'/../Fixtures');
         $loader = new PhpFileLoader(new ContainerBuilder(), new FileLocator($fixtures.'/config'));
 
-        $this->expectUserDeprecationMessageMatches('{^Since symfony/dependency-injection 7.4: Using \`\$this\` or its internal scope in config files is deprecated, use the \`\$loader\` variable instead in ".+" on line \d+\.$}');
+        $this->expectException(LogicException::class);
+        $this->expectExceptionMessageMatches('/^Using \`\$this\` or its internal scope in config files is not supported anymore, use the \`\$loader\` variable instead in ".+" on line \d+\.$/');
 
         $loader->load('legacy_internal_scope.php');
     }